the cache can now contains upto 0xFFFFFFFF items, added uwsgi.cache_exists to the api

+++ master.c
@@ -482,7 +482,7 @@ void master_loop(char **argv, char **environ) {
 
 			// remove expired cache items
 			if (uwsgi.cache_max_items > 0) {
-				for(i=0;i<uwsgi.cache_max_items;i++) {
+				for(i=0;i< (int)uwsgi.cache_max_items;i++) {
 					if (uwsgi.cache_items[i].expires) {
 						if (uwsgi.cache_items[i].expires < current_time) {
 							uwsgi_cache_del(uwsgi.cache_items[i].key, uwsgi.cache_items[i].keysize);
